organize header components
