+++ packages/blockchain-wallet-v4-frontend/src/components/WhatsNew/index.js
@@ -1,7 +1,6 @@
 import React from 'react'
 import { connect } from 'react-redux'
 import { selectors } from 'data'
-// import WhatsNewContent from './FaqContent'
 
 import WhatsNewLeora from './template.js'
 
@@ -11,7 +10,7 @@ class WhatsNewLeoraContainer extends React.Component {
   }
 
   render () {
-    const { WhatsNewContent, handleTrayRightToggle } = this.props
+    const { handleTrayRightToggle } = this.props
     return (
       <WhatsNewLeora handleTrayRightToggle={handleTrayRightToggle}/>
     )
