+++ packages/blockchain-wallet-v4-frontend/src/layouts/Wallet/Header/RefreshIcon/index.js
