--- packages/blockchain-wallet-v4-frontend/src/layouts/Wallet/Header/WhatsNew/selectors.js
@@ -1,3 +1,0 @@
-import { selectors } from 'data'
-
-export const getData = selectors.core.kvStore.whatsNew.getLastViewed
